requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.55
botocore==1.34.55
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
curl_cffi==0.6.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
docker==7.0.0
duckduckgo_search==4.5.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
groq==0.4.2
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.7.1
litellm==1.29.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==0.1.3
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.1.7
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
phidata==2.3.50
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.27.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
